requirements.txt
datasets>=2.14.0
openai>=1.0.0
python-dotenv>=1.0.0
